Fold Jaws benchmarks into jaws_test.go
